+++ packages/root/types/types/locationUtils.d.ts
@@ -1,4 +1,4 @@
-import { RawLocation, SessionLocation, PartialLocation, Key } from "./location";
+import { LocationComponents, SessionLocation, PartialLocation, Key } from "./location";
 export interface QueryFunctions {
     parse: (query?: string) => any;
     stringify: (query?: any) => string;
@@ -6,13 +6,12 @@ export interface QueryFunctions {
 export declare type RawPathname = (pathname: string) => string;
 export interface LocationUtilOptions {
     query?: QueryFunctions;
-    decode?: boolean;
     baseSegment?: string;
     raw?: RawPathname;
 }
 export interface LocationUtils {
-    keyed(location: RawLocation, key: Key): SessionLocation;
-    genericLocation(value: string | object, state?: any): RawLocation;
+    keyed(location: LocationComponents, key: Key): SessionLocation;
+    genericLocation(value: string | object, state?: any): LocationComponents;
     stringifyLocation(location: SessionLocation): string;
     stringifyLocation(location: PartialLocation): string;
 }